Moved the abstract runner method to its own file
